refactor: 로또 관련 메인 기능들 패키지 생성
