+++ src/main/java/lotto/domain/LottoGame.java
@@ -1,8 +1,9 @@
-package lotto;
+package lotto.domain;
 
 import camp.nextstep.edu.missionutils.Randoms;
 import java.util.List;
 import java.util.Optional;
+import lotto.InputHandler;
 import lotto.constant.Rank;
 
 public class LottoGame {
