+++ src/main/java/lotto/domain/WinnerResult.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.domain;
 
 import java.util.LinkedHashMap;
 import java.util.Map;
